Moved and renamed testing helpers
